refactor: extract event-dispatch.ts from daemon/main.ts
Move the POST /events route handler out of main() into a new
createEventDispatcher() factory in src/daemon/event-dispatch.ts,
reducing main.ts by ~250 lines. Wires planWatchConfig as a mutable
object reference for hot-reload support.
